Autocompleter for Principals (users and groups)
This is a combination of the autocompleter for users and groups. Both
object types will be searched, and their results will be displayed
alongside eachother. The return value is hardcoded: for users, return
their email address, for groups, return "Group:Name" (or if "Name"
matches any autocomplete delimiter, "Group:ID"). This corresponds with
what RT::Record::Role::Roles::_ResolveRoles accepts.

Autocompleters can now display an extra line of information (called
"detail"); for Principals it's used to display whether the record is a
user or group.