Handle Group:Name and Group:ID in role input fields
This was already implicitly available by providing a numeric id in the
role input field, but this capability makes it easier for end users.

This reuses _ResolveRoles in /Ticket/Create.html to scan for RT
addresses; without this change, anything other than email addresses
(specifically Group:Foo) will be stripped from the builtin role input
fields. (This check must be done on /Ticket/Create.html, not later, so
that we can cancel the form submission and let the user fix their
error.)