Cache OCFVs to improve performance searching for duplicates on add
When adding a new value to a custom field, the HasEntry code is used
to first determine whether the new value already exists. When inserting
a large number of values one after the other (like along list of IPs),
the HasEntry code pulled the full list of values from the DB for every
check, creating a performance issue. Cache the values to speed up the
existing value check.