Script to automatically update DB sequences to the next available value
When running a full serializer export and then an import using the
--clone option, the target database is typically initialized using
rt-setup-database --action create,schema,acl. This correctly creates
the empty tables, however Postgres and Oracle use sequences on RT's
tables and after the import, the sequences can still be at their initial
values. Provide a script to set sequences to the next available value.